[fix] we now live in the present
Dates from 1970 are now from 2018 forward
